+++ Helper/ScopeSetting.php
@@ -19,11 +19,13 @@ class ScopeSetting extends \Magento\Framework\App\Helper\AbstractHelper
     const CONSENT_AT_CHECKOUT_EMAIL_LIST_ID = 'klaviyo_reclaim_consent_at_checkout/email_consent/list_id';
     const CONSENT_AT_CHECKOUT_EMAIL_CONSENT_TEXT = 'klaviyo_reclaim_consent_at_checkout/email_consent/consent_text';
     const CONSENT_AT_CHECKOUT_EMAIL_SORT_ORDER = 'klaviyo_reclaim_consent_at_checkout/email_consent/sort_order';
+    const CONSENT_AT_CHECKOUT_EMAIL_LABEL_TEXT = 'klaviyo_reclaim_consent_at_checkout/email_consent/label_text';
 
     const CONSENT_AT_CHECKOUT_SMS_IS_ACTIVE = 'klaviyo_reclaim_consent_at_checkout/sms_consent/is_active';
     const CONSENT_AT_CHECKOUT_SMS_LIST_ID = 'klaviyo_reclaim_consent_at_checkout/sms_consent/list_id';
     const CONSENT_AT_CHECKOUT_SMS_CONSENT_TEXT = 'klaviyo_reclaim_consent_at_checkout/sms_consent/consent_text';
     const CONSENT_AT_CHECKOUT_SMS_SORT_ORDER = 'klaviyo_reclaim_consent_at_checkout/sms_consent/sort_order';
+    const CONSENT_AT_CHECKOUT_SMS_LABEL_TEXT = 'klaviyo_reclaim_consent_at_checkout/sms_consent/label_text';
 
     const KLAVIYO_NAME_DEFAULT = 'klaviyo';
     const KLAVIYO_USERNAME = 'klaviyo_reclaim_user/klaviyo_user/username';
@@ -234,6 +236,10 @@ class ScopeSetting extends \Magento\Framework\App\Helper\AbstractHelper
     {
         return $this->getScopeSetting(self::CONSENT_AT_CHECKOUT_EMAIL_SORT_ORDER);
     }
+    public function getConsentAtCheckoutEmailConsentLabelText()
+    {
+        return $this->getScopeSetting(self::CONSENT_AT_CHECKOUT_EMAIL_LABEL_TEXT);
+    }
 
 
     public function getConsentAtCheckoutSMSIsActive()
@@ -251,6 +257,10 @@ class ScopeSetting extends \Magento\Framework\App\Helper\AbstractHelper
     public function getConsentAtCheckoutSMSConsentSortOrder()
     {
         return $this->getScopeSetting(self::CONSENT_AT_CHECKOUT_SMS_SORT_ORDER);
+    }
+    public function getConsentAtCheckoutSMSConsentLabelText()
+    {
+        return $this->getScopeSetting(self::CONSENT_AT_CHECKOUT_SMS_LABEL_TEXT);
     }